# perf(types): 3x speedup MakePartSet