--- a/.changelog/unreleased/improvements/3117-significantly-speedup-make-partset.md
+++ b/.changelog/unreleased/improvements/3117-significantly-speedup-make-partset.md
@@ -1,0 +1,2 @@
+- [`types`] Significantly speedup types.MakePartSet, which is used in creating a block proposal
+  ([\#3117](https://github.com/cometbft/cometbft/issues/3117)