--- a/crypto/merkle/bench_test.go
+++ b/crypto/merkle/bench_test.go
@@ -40,3 +40,25 @@ func BenchmarkInnerHash(b *testing.B) {
 		b.Fatal("Benchmark did not run!")
 	}
 }
+
+// Benchmark the time it takes to hash a 64kb leaf, which is the size of
+// a block part.
+// This helps determine whether its worth parallelizing this hash for the propose.r
+func BenchmarkLeafHash64kb(b *testing.B) {
+	b.ReportAllocs()
+	leaf := make([]byte, 64*1024)
+	hash := sha256.New()
+
+	for i := 0; i < b.N; i++ {
+		leaf[0] = byte(i)
+		got := leafHashOpt(hash, leaf)
+		if g, w := len(got), sha256.Size; g != w {
+			b.Fatalf("size discrepancy: got %d, want %d", g, w)
+		}
+		sink = got
+	}
+
+	if sink == nil {
+		b.Fatal("Benchmark did not run!")
+	}
+}